[fe#37] rename: productCard, category 폴더 구조 세분화
